// src/app/interfaces/heroes.interface.ts
export interface Heroe {
  _id?: string;
  nombre: string;
  bio: string;
  aparicion: string;
  casa: string;
  img: string[] | string;
}


//Generado por la IA, pasandole el JSON  
export  interface Personaje {
    _id: string;
    nombre: string;
    bio: string;
    img: string[] | string; 
    aparicion: string; // ISO date string
    casa: string;
  }